Module seratodj::tag::markers2

The Serato Markers2 tag stores various kinds of track "markers" like Cue Points, Saved Loops, Flips.

It also stores information about the tracks' color in the tracklist and if the track's beatgrid is locked.

Note that some of this information is also stored in Serato Markers_, and Serato will prefer that data over the information stored in Serato Markers2 if it is present.

The minimum length of this tag seems to be 470 bytes, and shorter contents are padded with null bytes.

Structs

BPMLockMarker
A BPMLOCK marker.

Markers2
Represents the Serato Markers2 tag.

Markers2Content
Represents the base64-encoded content of the Serato Markers2 tag.

TrackColorMarker
A COLOR marker.

UnknownMarker
An unknown marker that we don't have a parser for.

Enums

Marker
A marker in the Serato Markers2 tag.

Functions

peek_newline_or_nullbyte
Returns a nullbyte or newline character from the input slice (non-consuming).

peek_nullbyte
Returns a nullbyte from the input slice (non-consuming).